Higher learning had me starved!
Sitting in prayer to the Nazareth Revolutionary. Our Lord and Savior, Christ Jesus, came to me and quoted scripture

Matt 4:3 And when the tempter came to him, he said, If thou be the Son of God, command that these stones be made bread. 4But he answered and said, It is written, Man shall not live by bread alone, but by every word that proceedeth out of the mouth of God.